upgrade 1.0.11

+++ trace/module/goods/respository/respository.go
@@ -1,8 +1,8 @@
 package respository
 
 type Test struct {
-	ID      uint   `gorm:"column:id" json:"id"`
-	GoodsID uint64 `gorm:"column:goods_id" json:"goods_id"`
+	ID      int64  `gorm:"column:id" json:"id"`
+	GoodsID int64  `gorm:"column:goods_id" json:"goods_id"`
 	Name    string `gorm:"column:name" json:"name"`
 }
